requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
black==24.10.0
blinker==1.4
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.4
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
environs==14.1.0
executing==2.2.0
GDAL==3.4.0
geographiclib==2.0
geopy==2.4.1
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
jeepney==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
marshmallow==3.26.0
matplotlib-inline==0.1.7
more-itertools==8.10.0
mypy-extensions==1.0.0
numpy==1.21.5
oauthlib==3.2.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.3.4
python-apt==2.4.0+ubuntu4
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
SecretStorage==3.3.1
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
stripe==11.4.1
traitlets==5.14.3
typing_extensions==4.12.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.3.0
wadllib==1.3.6
wcwidth==0.2.13
zipp==1.0.0
